1.3
Overhaul to bring this script in line with my other Robocopy Backup script available here:
https://github.com/Silvalined/Robocopy-Backup
Use of expression building technique via $RobocopyCommand variable to overhaul and simplify logic.
Added logic to Set-Credential so that this can run on PS5 minimum instead of PS7. 
The script now changes the power setting so the machine doesnt go to sleep (and changes back after).